add: sns data file so that it will be easier to maintain the data if sns link is increased
